Moved registry to another file. Simplified imports
