+++ app/assets/javascripts/discourse/app/lib/transformer/registry.js
@@ -1,0 +1,5 @@
+export const VALUE_TRANSFORMERS = Object.freeze([
+  // use only lowercase names
+  "header-notifications-avatar-size",
+  "home-logo-href",
+]);
