+++ app/assets/javascripts/discourse/tests/helpers/qunit-helpers.js
@@ -45,7 +45,6 @@ import { resetMentions } from "discourse/lib/link-mentions";
 import { forceMobile, resetMobile } from "discourse/lib/mobile";
 import { resetModelTransformers } from "discourse/lib/model-transformers";
 import { resetNotificationTypeRenderers } from "discourse/lib/notification-types-manager";
-import { resetTransformers } from "discourse/lib/plugin-api/transformer";
 import {
   clearCache as clearOutletCache,
   resetExtraClasses,
@@ -66,6 +65,7 @@ import {
   resetHighestReadCache,
   setTopicList,
 } from "discourse/lib/topic-list-tracker";
+import { resetTransformers } from "discourse/lib/transformer";
 import { clearRewrites } from "discourse/lib/url";
 import { resetUserMenuTabs } from "discourse/lib/user-menu/tab";
 import {
